Refactor create book to create content
