+++ src/app.rb
@@ -4,7 +4,7 @@ require_relative './person'
 require_relative './rental'
 require_relative './student'
 require_relative './teacher'
-
+require_relative './create_contents/create_book'
 class App
   def initialize(menu)
     @menu = menu
@@ -37,6 +37,22 @@ class App
     @menu.option_list
   end
 
+  def list_rentals_by_person_id
+    puts 'No person entry found' if @people.empty?
+    puts 'Enter person ID: '
+    id = gets.chomp.to_i
+    @people.each do |person|
+      next unless person.id == id
+
+      puts 'Rentals:'
+      person.rentals.each do |rental|
+        puts "Date: #{rental.date}, Book: \"#{rental.book.title}\" by #{rental.book.author} "
+      end
+    end
+    sleep 0.75
+    @menu.option_list
+  end
+
   def create_person
     puts 'Do you want to create a student (1) or a teacher (2)? [Input the number]: '
     type = gets.chomp.to_i
@@ -44,6 +60,30 @@ class App
     puts 'Invalid input' if type != 1 && type != 2
     add_student if type == 1
     add_teacher if type == 2
+    sleep 0.75
+    @menu.option_list
+  end
+
+  def create_rental
+    puts 'Select a book from the following list by number'
+    @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
+    book_id = gets.chomp.to_i
+    puts 'Invalid input!' if !(book_id.is_a? Integer) && book_id > @books.length
+
+    puts 'Select a person from the following list by number (not id)'
+    @people.each_with_index do |person, index|
+      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
+    end
+    person_id = gets.chomp.to_i
+    puts 'Invalide input!' if !(person_id.is_a? Integer) && person_id > @people.length
+
+    print 'Please enter the date in this format [yyyy-mm-dd]: '
+    date = gets.chomp.to_s
+
+    rental = Rental.new(date, @books[book_id], @people[person_id])
+    @rentals << rental
+
+    puts 'Rental created successfully'
     sleep 0.75
     @menu.option_list
   end
@@ -74,57 +114,6 @@ class App
     puts "Teacher with id #{teacher.id} successfully created"
   end
 
-  def create_book
-    puts 'What is the title of the book: '
-    title = gets.chomp
-    puts 'What is the author of the book: '
-    author = gets.chomp
-    book = Book.new(title, author)
-    @books << book
-    puts 'Book created successfully'
-    sleep 0.75
-    @menu.option_list
-  end
-
-  def create_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index { |book, index| puts "#{index}) Title: #{book.title}, Author: #{book.author}" }
-    book_id = gets.chomp.to_i
-    puts 'Invalid input!' if !(book_id.is_a? Integer) && book_id > @books.length
-
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
-      puts "#{index}) [#{person.class}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    person_id = gets.chomp.to_i
-    puts 'Invalide input!' if !(person_id.is_a? Integer) && person_id > @people.length
-
-    print 'Please enter the date in this format [yyyy-mm-dd]: '
-    date = gets.chomp.to_s
-
-    rental = Rental.new(date, @books[book_id], @people[person_id])
-    @rentals << rental
-
-    puts 'Rental created successfully'
-    sleep 0.75
-    @menu.option_list
-  end
-
-  def list_rentals_by_person_id
-    puts 'No person entry found' if @people.empty?
-    puts 'Enter person ID: '
-    id = gets.chomp.to_i
-    @people.each do |person|
-      next unless person.id == id
-
-      puts 'Rentals:'
-      person.rentals.each do |rental|
-        puts "Date: #{rental.date}, Book: \"#{rental.book.title}\" by #{rental.book.author} "
-      end
-    end
-    sleep 0.75
-    @menu.option_list
-  end
 
   def exit
     puts 'Goodbye...'
